requirements.txt
flask
flask-sqlalchemy
flask-login
psycopg2-binary
gunicorn
werkzeug